UI chnages in group + profile + other chnages

+++ src/components/providers/LoginProvider.tsx
@@ -203,6 +203,8 @@ export const LoginProvider: React.FC<LoginProviderProps> = ({ children }) => {
     }, [openLogin]);
 
     const updateLastStepOfCompleteProfile = useCallback((step: number) => {
+        const existingStep = StorageUtils.getItem<number>(STORAGE_KEYS.COMPLETE_PROFILE_STEP) || 1;
+        if(step <= existingStep) return;
         StorageUtils.setItem<number>(STORAGE_KEYS.COMPLETE_PROFILE_STEP, step);
         console.log("Updated COMPLETE_PROFILE_STEP in storage to:", step);
         setLastStepOfCompleteProfile(step);
